[NFCI][Analysis] Extract alloca slice analysis from SROA
Extract the memory slice analysis infrastructure from SROA into a new
MemSliceAnalysis library to enable future code reuse in other passes
(notably to replace the ad-hoc pointer slice analysis currently in SPIRV
and DXIL which relies heavily on GEP and alloca pointer types).
